+++ packages/analyzer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mahoraga-analyzer
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies
+  - mahoraga-core@0.5.1
+
 ## 0.3.3
 
 ### Patch Changes