+++ packages/sources/CHANGELOG.md
@@ -1,5 +1,12 @@
 # mahoraga-sources
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - mahoraga-core@0.5.1
+
 ## 0.2.3
 
 ### Patch Changes